Split threadtests.py into separate test-files
