+++ .github/workflows/run-tests.yml
@@ -43,4 +43,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Run the tests
       run: |
-        python -m unittest tests/threadtests.py
+        python -m unittest discover tests/
